+++ src/app/tools/zcash-payment-widget/PaymentRequestWidgetCodeSnippet.tsx
