+++ src/app/tools/zcash-payment-widget/config.ts
